Rename Formatter to ConsoleFormatter
This allows the `Airbrussh::Formatter` to once again be the primary entry point
to airbrussh.
