+++ lib/sshkit/formatter/airbrussh.rb
@@ -1,14 +1,12 @@
-require "airbrussh"
-require "airbrussh/delegating_formatter"
+require "airbrussh/formatter"
 
 # Capistrano's formatter configuration requires that the formatter class
-# be in the SSHKit::Formatter namespace.
+# be in the SSHKit::Formatter namespace. So we declare
+# SSHKit::Formatter::Airbrussh that simply functions as an alias for
+# Airbrussh::Formatter.
 module SSHKit
   module Formatter
-    class Airbrussh < Airbrussh::DelegatingFormatter
-      def initialize(io, config=::Airbrussh.configuration)
-        super(config.formatters(io))
-      end
+    class Airbrussh < Airbrussh::Formatter
     end
   end
 end
