+++ test/sshkit/formatter/airbrussh_test.rb
@@ -1,13 +1,13 @@
 require "minitest_helper"
 require "sshkit/formatter/airbrussh"
+require_relative "../../airbrussh/formatter_test.rb"
 
-class SSHKit::Formatter::AirbrusshTest < Minitest::Test
-  def test_uses_formatters_from_configuration_object
-    io = stub
-    config = Airbrussh::Configuration.new
-    config.expects(:formatters).with(io).returns([:fmt1, :fmt2])
+# SSHKit::Formatter::Airbrussh should behave identically to
+# Airbrussh::Formatter, so just reuse the Airbrussh::FormatterTest.
+class SSHKit::Formatter::AirbrusshTest < Airbrussh::FormatterTest
+  private
 
-    formatter = SSHKit::Formatter::Airbrussh.new(io, config)
-    assert_equal([:fmt1, :fmt2], formatter.formatters)
+  def formatter_class
+    SSHKit::Formatter::Airbrussh
   end
 end
